don't use a sneaky bitwise operator to check for the existence of a substring
The ~ operator in JavaScript is the bitwise NOT operator. It inverts the bits of its operand.

But in JavaScript, it's sometimes used in a "trick" to check for the existence of a substring or an element in an array.

Here's why:

The indexOf method returns the index of the first occurrence of the specified value in a string, or -1 if the value is not found. When ~ is applied to this result, it inverts the bits.

In JavaScript, -1 is represented as all 1s in binary (11111111 for an 8-bit binary number). When you apply the ~ operator to -1, you get 0, because it inverts all the bits and you get 00000000.

So, ~indexOf() will return a truthy value (anything but 0) if the value is found, and 0 (which is falsy) if the value is not found.